requirements.txt
flask
flask_sqlalchemy
flask_bcrypt
flask_jwt_extended
mysqlclient
python-dotenv
pytest
pytest-flask
tox
pymysql